Organized class files and libraries into lib folder
