+++ lib/channels.js
