+++ lib/emojis.js
