+++ lib/storable.js
